Completed Visual Servoing Project
