--- rb5_control/src/mpi_visual_control.py
@@ -1,387 +1,0 @@
-#!/usr/bin/env python
-import rospy
-import numpy as np
-from pid_controller import PIDcontroller, genTwistMsg, coord
-from tf.transformations import euler_from_quaternion
-import time
-from geometry_msgs.msg import Twist, PoseArray
-# from april_detection.msg import AprilTagDetectionArray
-from single_pid_controller import SinglePIDController
-from sensor_msgs.msg import Joy
-
-# Define global variables: open loop and closed loop status
-openLoop_Status = False
-closeLoop_Status = False
-wp_idx = 0
-waypoint_apriltag = np.array([[0.59,0.0,0.0], 
-                              [0.50,0.0,0.0],
-                              [1.10,0.0,0.0]])
-
-def gen_cmd_msg(update_value):
-    joy_msg = Joy()
-    joy_msg.axes = [0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0]
-    joy_msg.buttons = [0, 0, 0, 0, 0, 0, 0, 0]
-    update_move, update_slide, update_rotation = update_value
-    joy_msg.axes[1] = -update_move 
-    joy_msg.axes[0] = update_slide 
-    joy_msg.axes[2] = -update_rotation 
-    return joy_msg
-
-def gen_rot_msg():
-    joy_msg = Joy()
-    joy_msg.axes = [0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0]
-    joy_msg.buttons = [0, 0, 0, 0, 0, 0, 0, 0]
-    joy_msg.axes[2] = 0.3
-    return joy_msg
-
-def empty_joy_msg():
-    joy_msg = Joy()
-    joy_msg.axes = [0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0]
-    # joy_msg.buttons = [0, 0, 0, 0, 1, 0, 0, 0]
-    joy_msg.buttons = [0, 0, 0, 0, 0, 0, 0, 0]
-    return joy_msg
-
-# Callback function for visual closed loop control
-def vehiclePose_callback(aprilTagPose):
-    global openLoop_Status
-    global closeLoop_Status
-    global waypoint_apriltag
-    global wp_idx
-
-    if wp_idx == 3:
-        openLoop_Status = False
-
-    if openLoop_Status == True:
-        if len(aprilTagPose.poses) != 0:
-        #     rospy.loginfo("Searching for April Tag...")
-        #     set_twist = np.array([0.0, 0.0, 0.04])
-        #     msg_twist = genTwistMsg(set_twist)
-        #     pub_twist.publish(msg_twist)
-        #     rospy.sleep(0.5)
-        #     return
-        # else:
-
-            curr_Pose = aprilTagPose.poses[0].position
-            curr_Quat = aprilTagPose.poses[0].orientation
-
-            curr_PosX = curr_Pose.x
-            curr_PosZ = curr_Pose.z
-            _, curr_Agl, _ = euler_from_quaternion([curr_Quat.w,
-                                                    curr_Quat.x,
-                                                    curr_Quat.y,
-                                                    curr_Quat.z])
-            # current_state = np.array([curr_PosZ, curr_PosX, curr_Agl])
-
-            pid_x = SinglePIDController(0.8, 1, 1, 1.4, 0.5)
-            pid_x.setTarget(waypoint_apriltag[wp_idx][1])
-            error_x = pid_x.getError(curr_PosX)
-
-            pid_z = SinglePIDController(1, 5, 1.0, 0.4, 0.4)
-            pid_z.setTarget(waypoint_apriltag[wp_idx][0])
-            error_z = pid_z.getError(curr_PosZ)
-
-            pid_r = SinglePIDController(0.1, 0.5, 0.5, 0.5, 0.4, True)
-            pid_r.setTarget(waypoint_apriltag[wp_idx][2])
-            error_r = pid_r.getError(curr_Agl)
-
-            if np.abs(error_r) > 0.10:
-                update_value_r = pid_r.update(error_r)
-                joy_msg = gen_cmd_msg([0, 0, update_value_r])
-                rospy.loginfo("Modifying starting direction")
-                rospy.sleep(0.01)
-                joy_pub.publish(joy_msg)
-            elif np.abs(error_x) < 0.10 and np.abs(error_z) < 0.10 and np.abs(error_r) <= 0.10:
-                joy_pub.publish(empty_joy_msg())
-                rospy.loginfo("Position Reached")
-                rospy.sleep(0.01)
-                closeLoop_Status = True
-                wp_idx += 1
-                rospy.loginfo('Alignment Complete.')
-            elif np.abs(error_x) < 0.10 and np.abs(error_z) < 0.10:
-                update_value_r = pid_r.update(error_r)
-                joy_msg = gen_cmd_msg([0, 0, update_value_r])
-                rospy.loginfo("Modifying direction")
-                joy_pub.publish(joy_msg)
-                rospy.sleep(0.01)
-            else:
-                rospy.loginfo("Current error is: ")
-                print (error_z, error_x, error_r)
-                update_value_z = pid_z.update(curr_PosZ)
-                update_value_x = pid_x.update(curr_PosX)
-                update_value_r = pid_r.update(curr_Agl)
-                joy_msg = gen_cmd_msg([update_value_z, update_value_x, 0]) # debug
-                joy_pub.publish(joy_msg)
-                rospy.sleep(0.01)
-            
-            # # init pid controller
-            # pid = PIDcontroller(0.02,0.005,0.005)
-            # pid.setTarget(waypoint_apriltag[wp_idx])
-            # update_value = pid.update(current_state)
-            # pub_twist.publish(genTwistMsg(coord(update_value, current_state)))
-            # time.sleep(0.05)
-            # print(np.linalg.norm(pid.getError(current_state, waypoint_apriltag[wp_idx])))
-
-            # if (np.linalg.norm(pid.getError(current_state, waypoint_apriltag[wp_idx])) > 0.10): # check the error between current state and current way point
-            #     # calculate the current twist
-            #     update_value = pid.update(current_state)
-            #     # publish the twist
-            #     pub_twist.publish(genTwistMsg(coord(update_value, current_state)))
-            #     #print(coord(update_value, current_state))
-            #     time.sleep(0.05)
-            #     # update the current state
-            #     current_state += update_value
-            # elif (np.linalg.norm(pid.getError(current_state, waypoint_apriltag[wp_idx])) <= 0.10):
-            #     closeLoop_Status = True
-            #     wp_idx += 1
-            #     rospy.loginfo('Alignment Complete.')
-
-
-if __name__ == "__main__":
-
-    rospy.init_node("vision_control")
-    pub_twist = rospy.Publisher("/twist", Twist, queue_size=1)
-    joy_pub = rospy.Publisher('/joy', Joy, queue_size = 1)
-    sub_aprilTagPose = rospy.Subscriber('/april_poses', PoseArray, callback=vehiclePose_callback)
-    rospy.sleep(0.5)
-
-    # Set target waypoints (x, y, orientation)
-    waypoint = np.array([[0.0,0.0,0.0], 
-                         [0.5,0.0,0.0],
-                         [0.5,1.0,np.pi],
-                         [0.0,0.0,0.0]])
-
-    # init pid controller
-    pid = PIDcontroller(0.02,0.005,0.005)
-
-    # init current state
-    current_state = np.array([0.0,0.0,0.0])
-
-    # in this loop we will go through each way point.
-    # once error between the current state and the current way point is small enough, 
-    # the current way point will be updated with a new point.
-    for idx, wp in enumerate(waypoint):
-
-        # Skip the first waypoint (assume it is properly aligned)
-        if idx == 0: continue
-        print("move to way point", wp)
-
-        # open loop and closed loop status
-        openLoop_Status = False
-        closeLoop_Status = False
-
-        # set wp as the target point
-        pid.setTarget(wp)
-        # calculate the current twist
-        update_value = pid.update(current_state)
-        # publish the twist
-        pub_twist.publish(genTwistMsg(coord(update_value, current_state)))
-        #print(coord(update_value, current_state))
-        time.sleep(0.05)
-        # update the current state
-        current_state += update_value
-        # Open loop control while moving from one way point to another waypoint
-        while(np.linalg.norm(pid.getError(current_state, wp)) > 0.05): # check the error between current state and current way point
-            # calculate the current twist
-            update_value = pid.update(current_state)
-            # publish the twist
-            pub_twist.publish(genTwistMsg(coord(update_value, current_state)))
-            #print(coord(update_value, current_state))
-            time.sleep(0.05)
-            # update the current state
-            current_state += update_value
-
-        openLoop_Status=True
-
-        # Close loop control to realign itself before moving to the next waypoint
-        # After open loop control, april tag is assume to be within view of sight of the robot.
-        while (openLoop_Status==True and closeLoop_Status==False):
-            #rospy.loginfo("Self-aligning...")
-            pass
-                
-    # stop the car and exit
-    pub_twist.publish(genTwistMsg(np.array([0.0,0.0,0.0])))
-    openLoop_Status =False
-
-
-#!/usr/bin/env python
-import rospy
-import numpy as np
-import time
-from tf.transformations import euler_from_quaternion
-## Controller Libraries
-from pid_controller import PIDcontroller, genTwistMsg, coord
-from single_pid_controller import SinglePIDController
-from mpi_navigator import MegaNavigatorNode
-## Messages Libraries
-from sensor_msgs.msg import Joy
-from april_detection.msg import AprilTagDetectionArray
-from geometry_msgs.msg import Twist, PoseArray
-
-# Define global variables:
-openLoop_Status = False                         # Open Loop Completion Status
-closeLoop_Status = False                        # Closed Loop Completion Status
-wp_apt_idx = 0                                  # Waypoint index for Apriltag
-waypoints_apriltag = np.array([[0.59,0.0,0.0],   # Each Waypoint based on its nearest
-                              [0.50,0.0,0.0],   # AprilTag reference frame
-                              [1.10,0.0,0.0]])
-
-# Define global functions for the callbacks
-def gen_cmd_msg(cmd_vals):
-    """
-    Generate Joy Command
-    """
-    joy_cmd = Joy()
-    joy_cmd.axes = [0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0 ,0.0]
-    joy_cmd.buttons = [0, 0, 0, 0, 0, 0, 0, 0]
-    drive_cmd, slide_cmd, rotate_cmd = cmd_vals
-    joy_cmd.axes[1] = -drive_cmd                # +ve forward
-    joy_cmd.axes[0] = slide_cmd                 # +ve left
-    joy_cmd.axes[2] = -rotate_cmd               # +ve ccw
-    return joy_cmd
-
-def visualControl_callback(aprilTagPose):
-    """
-    Visual Closed Loop Control to realign robot pose.
-    """
-    global openLoop_Status
-    global closeLoop_Status
-    global waypoints_apriltag
-    global wp_apt_idx
-
-    if wp_apt_idx == len(waypoints_apriltag):
-        # Make sure does execute the visual control after completed
-        openLoop_Status = False
-    
-    if openLoop_Status == True and closeLoop_Status == False:
-        if len(aprilTagPose.detections) == 0:
-            rospy.loginfo("Searching for April Tag..." + str(aprilTagPose.header.seq))
-            set_twist = np.array([0.0, 0.0, 0.04])
-            msg_twist = genTwistMsg(set_twist)
-            pub_Twist.publish(msg_twist)
-            #rospy.sleep(0.05)
-        else:
-        # If april tag is detected
-        # if len(aprilTagPose.detections) != 0:
-
-            # Get April Tage Pose
-            curr_Pose = aprilTagPose.detections[0].pose.position
-            curr_Quat = aprilTagPose.detections[0].pose.orientation
-
-            curr_PosZ, curr_PosX = curr_Pose.z, curr_Pose.x
-            _, curr_Agl, _ = euler_from_quaternion([curr_Quat.w,
-                                                    curr_Quat.x,
-                                                    curr_Quat.y,
-                                                    curr_Quat.z])
-            
-            # This is actually x-axis map frame
-            pid_z = SinglePIDController(1.0, 5.0, 1.0, 1.0, 0.5)
-            pid_z.setTarget(waypoints_apriltag[wp_apt_idx][0])
-            error_z = pid_z.getError(curr_PosZ)
-            # This is actually y-axis map frame
-            pid_x = SinglePIDController(1.0, 1.0, 1.0, 1.0, 0.5)
-            pid_x.setTarget(waypoints_apriltag[wp_apt_idx][1])
-            error_x = pid_x.getError(curr_PosX)
-            # Orientation
-            pid_r = SinglePIDController(0.1, 0.5, 0.5, 0.5, 0.4, True)
-            pid_r.setTarget(waypoints_apriltag[wp_apt_idx][2])
-            error_r = pid_r.getError(curr_Agl)
-
-            # Check waypoint reached
-            if np.abs(error_x) < 0.08 and np.abs(error_z) < 0.08 and np.abs(error_r) < 0.25:
-                pub_Joy.publish(gen_cmd_msg([0.0, 0.0, 0.0]))
-                rospy.sleep(0.05)
-                rospy.loginfo("Waypoint Reached and Alignment Complete")
-                closeLoop_Status = True
-                wp_apt_idx += 1
-            # Check heading first
-            elif np.abs(error_r) >= 0.25:         # About 15 Deg.
-                update_value_r = pid_r.update(error_r)
-                joy_cmd = gen_cmd_msg([0, 0, update_value_r])
-                rospy.loginfo("Adjusting Heading")
-                pub_Joy.publish(joy_cmd)
-                rospy.sleep(0.05)
-            else:
-                rospy.loginfo("Current error is: ")
-                print (error_z, error_x, error_r)
-                update_value_z = pid_z.update(curr_PosZ)
-                update_value_x = pid_x.update(curr_PosX)
-                update_value_r = pid_r.update(curr_Agl)
-                joy_cmd = gen_cmd_msg([update_value_z, update_value_x, 0]) # debug
-                pub_Joy.publish(joy_cmd)
-                rospy.sleep(0.05)
-
-
-if __name__ == "__main__":
-
-    # Initialize Node
-    rospy.init_node("vision_control")
-
-    # Publishers
-    pub_Twist = rospy.Publisher("/twist", Twist, queue_size = 1)
-    pub_Joy = rospy.Publisher('/joy', Joy, queue_size = 1)
-    # Subscriber
-    sub_aprilTagPose = rospy.Subscriber('/apriltag_detection_array', AprilTagDetectionArray, callback=visualControl_callback)
-
-    # Set target waypoints (x, y, orientation)
-    waypoints = np.array([[0.0,0.0,0.0], 
-                         [0.5,0.0,0.0],
-                         [0.5,1.0,np.pi],
-                         [0.0,0.0,0.0]])
-
-    # Initialize PID Controller
-    pid = PIDcontroller(0.02,0.01,0.01)
-
-    # Initialize the current state
-    current_state = waypoints[0,:]
-
-    # Loop through waypoints
-    for idx, wp in enumerate(waypoints):
-
-        # Skip the first waypoint (assume it is properly aligned)
-        if idx == 0: continue
-
-        # Show next waypoint log
-        rospy.loginfo("Next Waypoint ( %s m, %s m, %s rad )", wp[0], wp[1], wp[2])
-
-        # Set Next Waypoint
-        pid.setTarget(wp)
-
-        # Calculate the current twist
-        update_value = pid.update(current_state)
-
-        # Publish the twist
-        pub_Twist.publish(genTwistMsg(coord(update_value, current_state)))
-        time.sleep(0.05)
-
-        # Update the current state
-        current_state += update_value
-
-        # Open loop control while moving from one way point to another waypoint
-        while(np.linalg.norm(pid.getError(current_state, wp)) > 0.05): # check the error between current state and current way point
-            
-            # Calculate the current twist
-            update_value = pid.update(current_state)
-
-            # Publish the twist
-            pub_Twist.publish(genTwistMsg(coord(update_value, current_state)))
-            time.sleep(0.05)
-
-            # Update the current state
-            current_state += update_value
-
-        # Mark open loop control complete
-        openLoop_Status = True
-
-        # Close loop control to realign itself before moving to the next waypoint
-        # After open loop control, april tag is assume to be within view of sight of the robot.
-        while (openLoop_Status==True and closeLoop_Status==False):
-            # Execute in callback function
-            pass
-
-        # Reset loop status
-        closeLoop_Status = False
-        openLoop_Status = False
-
-    # Stop the car and exit
-    pub_Twist.publish(genTwistMsg(np.array([0.0,0.0,0.0])))
-    openLoop_Status =False
