+++ telemetry_data/video_path.csv
